refactor to allow for command line options for terraform provider and version selection
