+++ openapi-to-terraform.Generator/ITerraformGenerator.cs
@@ -1,0 +1,8 @@
+using Microsoft.OpenApi.Models;
+
+namespace openapi_to_terraform.Generator {
+    interface ITerraformGenerator {
+        void GenerateWithTerraformVars(OpenApiDocument document);
+        void GenerateWithTemplateFiles(OpenApiDocument document);
+    }
+}
