+++ openapi-to-terraform.Generator/azurerm/v2_45_1/GeneratorModels/TerraformApimOperation.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Text;
 using Microsoft.OpenApi.Models;
 
-namespace openapi_to_terraform.Generator.GeneratorModels
+namespace openapi_to_terraform.Generator.azurerm.v2_45_1.GeneratorModels
 {
     public class TerraformApimOperation
     {
