+++ openapi-to-terraform.Generator/azurerm/v2_45_1/Generators/ApiGenerator.cs
@@ -1,9 +1,9 @@
 using System.Collections.Generic;
 using System.Text;
 using Microsoft.OpenApi.Models;
-using openapi_to_terraform.Generator.GeneratorModels;
+using openapi_to_terraform.Generator.azurerm.v2_45_1.GeneratorModels;
 
-namespace openapi_to_terraform.Generator.Generators
+namespace openapi_to_terraform.Generator.azurerm.v2_45_1.Generators
 {
     public class ApiGenerator
     {
