+++ openapi-to-terraform.Generator/azurerm/v2_45_1/TemplateAppliers/ApiTemplateApplier.cs
@@ -1,4 +1,4 @@
-namespace openapi_to_terraform.Generator.TemplateAppliers
+namespace openapi_to_terraform.Generator.azurerm.v2_45_1.TemplateAppliers
 {
     public class ApiTemplateApplier
     {
